requirements.txt
numpy==1.18.5
pandas==1.2.4
Pillow==8.4.0
scipy==1.4.1
streamlit==1.20.0
opencv-python==4.5.2.54
ipython==8.11.0
-e .